chapter-06/lab-6.3/requirements.txt
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
crowdstrike-falconpy==1.3.5
docutils==0.20.1
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
statistics==1.0.3.5
urllib3==2.1.0
